feat: native module register, painter
